Moved classes into folders
