+++ CMakeLists.txt
@@ -4,19 +4,19 @@ project(ConsoleGame)
 set(CMAKE_CXX_STANDARD 20)
 
 add_executable(ConsoleGame  main.cpp
-                            Unit.h
-                            Unit.cpp
-                            Enemy.h
-                            Enemy.cpp
-                            Protagonist.h
-                            Protagonist.cpp
-                            Pyromancer.h
-                            Pyromancer.cpp
-                            Imp.cpp
-                            Imp.h
-                            DarkMage.h
-                            DarkMage.cpp
-                            EnemyFactory.h
-                            EnemyFactory.cpp
-                            ImpFactory.h
-                            ImpFactory.cpp)
+        Units/Unit.h
+        Units/Unit.cpp
+        Units/Enemy.h
+        Units/Enemy.cpp
+        Units/Protagonist.h
+        Units/Protagonist.cpp
+        Units/Pyromancer.h
+        Units/Pyromancer.cpp
+        Units/Imp.cpp
+        Units/Imp.h
+        Units/DarkMage.h
+        Units/DarkMage.cpp
+        Factories/EnemyFactory.h
+        Factories/EnemyFactory.cpp
+        Factories/ImpFactory.h
+        Factories/ImpFactory.cpp)
